consumer2/pyproject.toml
[project]
name = "order-consumer"
version = "0.1.0"
description = "Order API Consumer for Pact Contract Testing POC"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.2",
    "pact-python>=3.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
